refactorings after merge
