+++ apps/server/src/modules/tool/service/index.ts
@@ -1,0 +1,9 @@
+export * from './validation/common-tool-validation.service';
+export * from './validation/external-tool-validation.service';
+export * from './validation/school-external-tool-validation.service';
+
+export * from './course-external-tool.service';
+export * from './external-tool-version.service';
+export * from './lti11.service';
+export * from './external-tool.service';
+export * from './school-external-tool.service';
