+++ apps/server/src/modules/tool/service/validation/external-tool-validation.service.spec.ts
@@ -4,12 +4,12 @@ import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { ExternalToolDO } from '@shared/domain/domainobject/external-tool';
 import { UnprocessableEntityException } from '@nestjs/common';
 import { ExternalToolService } from '../external-tool.service';
-import { ToolValidationService } from './tool-validation.service';
+import { ExternalToolValidationService } from './external-tool-validation.service';
 import { CommonToolValidationService } from './common-tool-validation.service';
 
 describe('ToolValidation', () => {
 	let module: TestingModule;
-	let service: ToolValidationService;
+	let service: ExternalToolValidationService;
 
 	let externalToolService: DeepMocked<ExternalToolService>;
 	let commonToolValidationService: DeepMocked<CommonToolValidationService>;
@@ -17,7 +17,7 @@ describe('ToolValidation', () => {
 	beforeAll(async () => {
 		module = await Test.createTestingModule({
 			providers: [
-				ToolValidationService,
+				ExternalToolValidationService,
 				{
 					provide: ExternalToolService,
 					useValue: createMock<ExternalToolService>(),
@@ -29,7 +29,7 @@ describe('ToolValidation', () => {
 			],
 		}).compile();
 
-		service = module.get(ToolValidationService);
+		service = module.get(ExternalToolValidationService);
 		externalToolService = module.get(ExternalToolService);
 		commonToolValidationService = module.get(CommonToolValidationService);
 	});
@@ -82,7 +82,7 @@ describe('ToolValidation', () => {
 				const result: Promise<void> = service.validateCreate(externalToolDO);
 
 				await expect(result).rejects.toThrow(
-					new UnprocessableEntityException(`The Client Id of the tool: ${externalToolDO.name} is already used`)
+					new UnprocessableEntityException(`The Client Id of the tool ${externalToolDO.name} is already used.`)
 				);
 			});
 		});
@@ -96,6 +96,7 @@ describe('ToolValidation', () => {
 		it('should call the common validation service', async () => {
 			const externalToolDO: ExternalToolDO = externalToolDOFactory.buildWithId();
 			externalToolDO.id = 'toolId';
+			externalToolService.isOauth2Config.mockReturnValue(false);
 
 			await service.validateUpdate(externalToolDO.id, externalToolDO);
 
@@ -105,6 +106,7 @@ describe('ToolValidation', () => {
 		describe('when checking if parameter id matches toolId', () => {
 			it('should throw an error if not matches', async () => {
 				const externalToolDO: ExternalToolDO = externalToolDOFactory.buildWithId();
+				externalToolService.isOauth2Config.mockReturnValue(true);
 
 				const func = () => service.validateUpdate('notMatchToolId', externalToolDO);
 
@@ -116,6 +118,7 @@ describe('ToolValidation', () => {
 			it('should return without error if matches', async () => {
 				const externalToolDO: ExternalToolDO = externalToolDOFactory.buildWithId();
 				externalToolDO.id = 'toolId';
+				externalToolService.isOauth2Config.mockReturnValue(false);
 
 				const result: Promise<void> = service.validateUpdate(externalToolDO.id, externalToolDO);
 
@@ -130,6 +133,7 @@ describe('ToolValidation', () => {
 					.withOauth2Config({ clientId: 'clientId1' })
 					.buildWithId();
 				externalToolService.findExternalToolById.mockResolvedValue(existingExternalToolDO);
+				externalToolService.isOauth2Config.mockReturnValue(true);
 
 				const result: Promise<void> = service.validateUpdate(externalToolDO.id as string, externalToolDO);
 
@@ -141,6 +145,7 @@ describe('ToolValidation', () => {
 			it('should pass if tool has the same clientId as before', async () => {
 				const externalToolDO: ExternalToolDO = externalToolDOFactory.withOauth2Config().buildWithId();
 				externalToolService.findExternalToolById.mockResolvedValue(externalToolDO);
+				externalToolService.isOauth2Config.mockReturnValue(true);
 
 				const result: Promise<void> = service.validateUpdate(externalToolDO.id as string, externalToolDO);
 
@@ -155,6 +160,7 @@ describe('ToolValidation', () => {
 					.withOauth2Config({ clientId: 'clientId1' })
 					.buildWithId();
 				externalToolService.findExternalToolById.mockResolvedValue(existingExternalToolDO);
+				externalToolService.isOauth2Config.mockReturnValue(true);
 
 				const result: Promise<void> = service.validateUpdate(externalToolDO.id as string, externalToolDO);
