+++ apps/server/src/modules/tool/service/validation/external-tool-validation.service.ts
@@ -4,7 +4,7 @@ import { ExternalToolService } from '../external-tool.service';
 import { CommonToolValidationService } from './common-tool-validation.service';
 
 @Injectable()
-export class ToolValidationService {
+export class ExternalToolValidationService {
 	constructor(
 		private readonly externalToolService: ExternalToolService,
 		private readonly commonToolValidationService: CommonToolValidationService
@@ -14,7 +14,7 @@ export class ToolValidationService {
 		await this.commonToolValidationService.validateCommon(externalToolDO);
 
 		if (externalToolDO.config instanceof Oauth2ToolConfigDO && !(await this.isClientIdUnique(externalToolDO))) {
-			throw new UnprocessableEntityException(`The Client Id of the tool: ${externalToolDO.name} is already used`);
+			throw new UnprocessableEntityException(`The Client Id of the tool ${externalToolDO.name} is already used.`);
 		}
 	}
 
@@ -35,7 +35,7 @@ export class ToolValidationService {
 
 		const loadedTool: ExternalToolDO = await this.externalToolService.findExternalToolById(toolId);
 		if (
-			loadedTool.config instanceof Oauth2ToolConfigDO &&
+			this.externalToolService.isOauth2Config(loadedTool.config) &&
 			externalToolDO.config &&
 			externalToolDO.config.type !== loadedTool.config.type
 		) {
@@ -43,8 +43,9 @@ export class ToolValidationService {
 		}
 
 		if (
-			externalToolDO.config instanceof Oauth2ToolConfigDO &&
-			loadedTool.config instanceof Oauth2ToolConfigDO &&
+			externalToolDO.config &&
+			this.externalToolService.isOauth2Config(externalToolDO.config) &&
+			this.externalToolService.isOauth2Config(loadedTool.config) &&
 			externalToolDO.config.clientId !== loadedTool.config.clientId
 		) {
 			throw new UnprocessableEntityException(`The Client Id of the tool ${externalToolDO.name || ''} is immutable.`);
