+++ apps/server/src/modules/tool/uc/external-tool.uc.spec.ts
@@ -12,8 +12,7 @@ import {
 	oauth2ToolConfigDOFactory,
 } from '@shared/testing/factory/domainobject/external-tool.factory';
 import { ExternalToolUc } from './external-tool.uc';
-import { ExternalToolService } from '../service/external-tool.service';
-import { ExternalToolValidationService } from '../service/validation/external-tool-validation.service';
+import { ExternalToolService, ExternalToolValidationService } from '../service';
 
 describe('ExternalToolUc', () => {
 	let module: TestingModule;
