+++ apps/server/src/modules/tool/uc/school-external-tool.uc.spec.ts
@@ -1,16 +1,14 @@
 import { createMock, DeepMocked } from '@golevelup/ts-jest';
 import { MikroORM } from '@mikro-orm/core';
 import { Test, TestingModule } from '@nestjs/testing';
-import { Actions, Permission, User, EntityId } from '@shared/domain';
+import { Actions, EntityId, Permission, User } from '@shared/domain';
 import { SchoolExternalToolDO } from '@shared/domain/domainobject/external-tool/school-external-tool.do';
 import { schoolExternalToolDOFactory } from '@shared/testing/factory/domainobject/school-external-tool.factory';
 import { setupEntities, userFactory } from '@shared/testing';
-import { AuthorizationService, AllowedAuthorizationEntityType } from '@src/modules/authorization';
-import { SchoolExternalToolService } from '../service/school-external-tool.service';
+import { AllowedAuthorizationEntityType, AuthorizationService } from '@src/modules/authorization';
+import { CourseExternalToolService, SchoolExternalToolService, SchoolExternalToolValidationService } from '../service';
 import { SchoolExternalToolQueryInput } from './dto/school-external-tool.types';
 import { SchoolExternalToolUc } from './school-external-tool.uc';
-import { CourseExternalToolService } from '../service/course-external-tool.service';
-import { SchoolExternalToolValidationService } from '../service/school-external-tool-validation.service';
 
 describe('SchoolExternalToolUc', () => {
 	let module: TestingModule;
